Left White Window Added Left Margin: 10px (Anchored to left) Width: 340px Top and Bottom Margin: 40px (Anchored to top and bottom)
Component used: BorderContainer
It will allow using child componens with the flexibility to lay out controls
using x y properties OR left, right, top and bottom anchor properties.

The minimum width of the application is set to 960. The Sales Info section can be
anchored to bottom and top. No need to specify height for Sales Info section.
Other sections should be top anchored with height specified.
When the component resizes, the Sales Info will resize except the rest sections.
If the Sales Info section is going to be a simple textual kind of conten,
then depending on the style of content, Label or TextArea controls can be used.
For TextArea the scrollbars can be hidden.