Right White Window Added Left Margin: 360px (Anchored to left) Right Margin: 10px (Anchored to right) Top and Bottom Margin: 40px (Anchored to top and bottom)
Component used:
BorderContainer as main container
Scroller for the scroll view
Skinned the Horizontal and Vertcal scroll bars to get the look as in AI

Restructured the packaging of skins:
The right and left window share same look and feel.
BasePanelSkin degines the common look.
RightPanelSkin and LeftPanelSkin are subclass of BasePanelSkin. Additional customization
if needed in future can be done in these classes.

The Scoller needs a viewport to hold the actual content which isa Group container.
Any content that needs to be shown in the right window should be a child
of the Scroller/Group.